[ARM] qsd8k: camera: use epoch1 interrupt to control the flash more precisely
We use the epoch1 interrupt to control flash timing in the kernel. The
purpose is to reduce flash duration as much as possible. Userspace
has no way to control the exact timing, but the VFE can do it. Flash should
cover the exposure period of each line. Currently we skip a frame during
snapshot, so snapshot uses two frames. We want to fire the flash in the
middle of the first frame when the exposure time is short. We are using
rolling shutter. It means flash duration = readout time + exposure time.
We have to fire the flash earlier to cover the exposure time of the first
line of the second frame.

The epoch1 interrupt allows us to set a line index and we will get an
interrupt when VFE reaches the line. VFE will fire the flash in high mode
when it gets the epoch1 interrupt. VFE will turn off the flash after
snapshot is done.

Signed-off-by: Iliyan Malchev <malchev@google.com>